OXDEV-470 Remove 'Your market' dropdown from setup
During the setup of OXID eShop, shop owners should not need to enter information, which is not used in the shop.
The value from the drop down 'Your market' is not needed any longer, as it belongs to the deprecated and unused dynPages funtionality.
The value has been removed from the demo data.